Add the ability for the caller to inject pre/post request functionality and functionlity to setup the requestLib.
This functionality is useful in some cases where the requestLib
needs to be setup to use insecure mode to do not CA signed SSL certs.
In addition, if not supporting expressed auth flows, and the target REST
API returns a fault code an addition pre-request operation can be injected to authenticate
the session based upon that or the organization (e.g. SSO, SAML, etc).  Given this
flow the request can be respun given the original request.